refactor(models/cca): inline ELIC-style codec wiring; move convert to examples
Apply the same Elic2022Official-style treatment as stf/tcm to the CCA
model:

- inline the main entropy stack as a CCAModel._build_main_latent_codec
  method (HyperpriorLatentCodec + ChannelGroupsLatentCodec dict
  comprehensions) and inline the aux branch's ChannelGroupsLatentCodec in
  _CCAAuxEntropyModel.__init__, replacing the build_channel_slice_codec /
  _build_cca_main_latent_codec / _make_cca_head / _cumsum_with_zero
  factory layer;
- drop the load_state_dict auto-convert override and the from_state_dict
  upstream-detection branch so the model only loads compressai-layout
  state dicts;
- move convert_upstream_cca_state_dict + helpers
  (_is_upstream_cca_state_dict, _find_naf_*_prefixes, _reroot_per_slice_keys,
  _replicate_gaussian_conditional, rename tables) to
  examples/convert_cca_checkpoint.py;
- keep _infer_config_from_state_dict in cca.py (used by from_state_dict);
- point test_cca_upstream_state_dict_conversion at the examples script via
  _load_convert_fn.

State_dict layout, forward numerics, and conversion output are unchanged
(verified key-set identical and max abs diff 0.0 across x_hat / y / z /
aux likelihoods vs. the pre-refactor model).
